requirements.txt
amqp==2.2.2
astroid==1.5.3
backports.functools-lru-cache==1.4
beautifulsoup4==4.6.0
billiard==3.5.0.3
boto==2.47.0
bs4==0.0.1
bz2file==0.98
captcha==0.2.4
celery==4.1.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
colorlog==3.1.2
configparser==3.5.0
cycler==0.10.0
enum34==1.1.6
Flask==0.12.2
funcsigs==1.0.2
gensim==2.1.0
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
h5py==2.7.1
idna==2.5
isort==4.2.15
itsdangerous==0.24
jieba==0.38
Jinja2==2.10
kombu==4.1.0
lazy-object-proxy==1.3.1
lightgbm==2.0.10
locustio==0.8.1
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
mock==2.0.0
msgpack-python==0.4.8
numpy==1.13.0
olefile==0.44
opencv-python==3.3.0.10
pandas==0.20.2
pbr==3.0.1
peewee==2.10.2
Pillow==5.0.0
progressbar==2.3
protobuf==3.3.0
py==1.5.2
pylint==1.7.2
pyparsing==2.2.0
pytesseract==0.1.7
python-dateutil==2.6.0
pytz==2017.2
pyzmq==16.0.3
redis==2.10.6
requests==2.17.3
scikit-learn==0.18.1
scipy==0.19.0
singledispatch==3.4.0.3
six==1.10.0
smart-open==1.5.3
subprocess32==3.2.7
tensorflow==1.1.0
urllib3==1.21.1
vine==1.1.4
Werkzeug==0.12.2
wrapt==1.10.11
xgboost==0.6
